상품가입하기 수정,dto 이름 통일

+++ src/main/java/com/gang/start/account/BankAccountDTO.java
@@ -9,7 +9,7 @@ public class BankAccountDTO {
 	
 	private Long  accountNum; 
 	private String userName;
-	private Integer bookNum;
+	private long bookNum;
 	private Date accountDate;
 	private BankBookDTO bankBookDTO;
 	
@@ -35,10 +35,10 @@ public class BankAccountDTO {
 	public void setUserName(String userName) {
 		this.userName = userName;
 	}
-	public Integer getBookNum() {
+	public long getBookNum() {
 		return bookNum;
 	}
-	public void setBookNum(Integer bookNum) {
+	public void setBookNum(long bookNum) {
 		this.bookNum = bookNum;
 	}
 	public Date getAccountDate() {
